chore: fix all clippy warnings
- Add clippy.toml to suppress needless_range_loop in SIMD/math code
- Add missing Default impls (AudioInput, AudioOutput, BackendRegistry,
  DryWetMix, GraphExecutor, MpscQueue, RingQueue, ReadHead, SimdDetector)
- Fix manual slice copy -> copy_from_slice (6 places)
- Fix clamp patterns (15 places)
- Fix identity map, matches! macro, unnecessary braces
- Fix doc comments (empty, orphaned, indentation)
- Add is_empty() to RingQueue
- Suppress should_implement_trait on from_str/next methods
- Remove unused lifetime in DspContext

+++ rill-sampler/src/wav.rs
@@ -47,7 +47,7 @@ pub fn load_wav(path: &str) -> Result<SampleBuffer<Sample>, WavError> {
     let mut reader = hound::WavReader::open(path)?;
     let spec = reader.spec();
 
-    let channels = spec.channels as u16;
+    let channels = spec.channels;
     let sample_rate = spec.sample_rate as f32;
     let bits_per_sample = spec.bits_per_sample;
